requirements.txt
Flask
playwright
pywebview
pyinstaller